Title: 길이 단위 환산하기, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 251124/길이 단위 환산하기/convert-length-units.cpp
@@ -4,12 +4,12 @@ using namespace std;
 int main() {
     // Please write your code here.
 
-    int n;
+    double n;
     cin >> n;
 
     cout << fixed;
     cout.precision(1);
-    cout << n *30.48;
+    cout << n * 30.48;
 
     return 0;
 }
